services/parser-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
requests==2.32.3
pypdf==5.4.0
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
extract-msg==0.52.0
mail-parser==4.1.2
beautifulsoup4==4.13.3
lxml==5.3.2
